auth-sanctum with vue

+++ routes/api.php
@@ -3,6 +3,7 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\PostController;
+use App\Http\Resources\User as UserResource;
 
 /*
 |--------------------------------------------------------------------------
@@ -14,14 +15,17 @@ use App\Http\Controllers\PostController;
 | is assigned the "api" middleware group. Enjoy building your API!
 |
 */
-Route::middleware('auth:api')->get('/user', function (Request $request) {
+/*Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
+});*/
+Route::middleware('auth:sanctum')->get('/user', function (Request $request){
+    return new UserResource($request->user());
 });
 //Post
 Route::get('/posts/categories', [PostController::class, 'categories']);
 Route::get('/posts', [PostController::class, 'index']);
 Route::get('/posts/{id}', [PostController::class, 'show']);
 Route::get('/posts/category/{name}', [PostController::class, 'category']);
-Route::post('/posts/store', [PostController::class, 'store']);
-Route::put('/posts/{id}', [PostController::class, 'update']);
-Route::delete('/posts/{id}', [PostController::class, 'delete']);
+Route::post('/posts/store', [PostController::class, 'store'])->middleware('auth:sanctum');
+Route::put('/posts/{id}', [PostController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('/posts/{id}', [PostController::class, 'delete'])->middleware('auth:sanctum');
